+++ R/NBDCtools-package.R
@@ -9,5 +9,6 @@
 #' @importFrom stats na.omit setNames
 #' @importFrom sjmisc is_empty
 #' @importFrom haven is_tagged_na
+#' @importFrom utils packageVersion tail
 ## usethis namespace: end
 NULL